+++ lib/src/force_type_base.dart
@@ -2,43 +2,18 @@
 //
 // Copyright (c) 2023 Zaki Mubarok (kakzaki.dev)
 
-extension DynamicCasting on dynamic {
-  /// forcing dynamic type to String
-  String get forceString => "${this ?? ""}";
-
-  /// forcing dynamic type to Integer
-  int get forceInteger =>
-      int.tryParse("${this ?? "0"}".replaceAll(".", "").replaceAll(",", "")) ??
-      0;
-
-  /// forcing dynamic type to DateTime
-  DateTime get forceDateTime =>
-      DateTime.tryParse("${this ?? DateTime.now()}") ?? DateTime.now();
-
-  /// forcing dynamic type to Double
-  double get forceDouble =>
-      double.tryParse("${this ?? "0"}".replaceAll(",", ".")) ?? 0;
-
-  /// forcing dynamic type to Boolean
-  bool get forceBoolean =>
-      "${this ?? ""}" == '1' || "${this ?? ""}".toLowerCase() == 'true'
-          ? true
-          : false;
-}
-
 extension StringCasting on String {
   /// forcing String type to String
   String get forceString => this;
 
   /// forcing String type to Integer
-  int get forceInteger =>
-      int.tryParse((this).replaceAll(".", "").replaceAll(",", "")) ?? 0;
+  int get forceInteger => int.tryParse(this) ?? 0;
 
   /// forcing String type to DateTime
   DateTime get forceDateTime => DateTime.tryParse(this) ?? DateTime.now();
 
   /// forcing String type to Double
-  double get forceDouble => double.tryParse((this).replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse((this)) ?? 0;
 
   /// forcing String type to Boolean
   bool get forceBoolean =>
@@ -50,16 +25,14 @@ extension NullableStringCasting on String? {
   String get forceString => this ?? "";
 
   /// forcing String? type to Integer
-  int get forceInteger =>
-      int.tryParse((this ?? "0").replaceAll(".", "").replaceAll(",", "")) ?? 0;
+  int get forceInteger => int.tryParse((this ?? "0")) ?? 0;
 
   /// forcing String? type to DateTime
   DateTime get forceDateTime =>
       DateTime.tryParse("${this ?? DateTime.now()}") ?? DateTime.now();
 
   /// forcing String? type to Double
-  double get forceDouble =>
-      double.tryParse((this ?? "0").replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse((this ?? "0")) ?? 0;
 
   /// forcing String? type to Boolean
   bool get forceBoolean =>
@@ -73,14 +46,13 @@ extension IntegerCasting on int {
   String get forceString => "$this";
 
   /// forcing Integer type to Integer
-  int get forceInteger =>
-      int.tryParse("$this".replaceAll(".", "").replaceAll(",", "")) ?? 0;
+  int get forceInteger => this;
 
   /// forcing Integer type to DateTime
   DateTime get forceDateTime => DateTime.tryParse("$this") ?? DateTime.now();
 
   /// forcing Integer type to Double
-  double get forceDouble => double.tryParse("$this".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("$this") ?? 0;
 
   /// forcing Integer type to Boolean
   bool get forceBoolean =>
@@ -92,17 +64,14 @@ extension NullableIntegerCasting on int? {
   String get forceString => "${this ?? ""}";
 
   /// forcing NullableInteger type to Integer
-  int get forceInteger =>
-      int.tryParse("${this ?? "0"}".replaceAll(".", "").replaceAll(",", "")) ??
-      0;
+  int get forceInteger => this ?? 0;
 
   /// forcing NullableInteger type to DateTime
   DateTime get forceDateTime =>
       DateTime.tryParse("${this ?? DateTime.now()}") ?? DateTime.now();
 
   /// forcing NullableInteger type to Double
-  double get forceDouble =>
-      double.tryParse("${this ?? "0"}".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("${this ?? "0"}") ?? 0;
 
   /// forcing NullableInteger type to Boolean
   bool get forceBoolean =>
@@ -116,14 +85,13 @@ extension DoubleCasting on double {
   String get forceString => "$this";
 
   /// forcing Double type to Integer
-  int get forceInteger =>
-      int.tryParse("$this".replaceAll(".", "").replaceAll(",", "")) ?? 0;
+  int get forceInteger => toInt();
 
   /// forcing Double type to DateTime
   DateTime get forceDateTime => DateTime.tryParse("$this") ?? DateTime.now();
 
   /// forcing Double type to Double
-  double get forceDouble => double.tryParse("$this".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("$this") ?? 0;
 
   /// forcing Double type to Boolean
   bool get forceBoolean =>
@@ -135,17 +103,14 @@ extension NullableDoubleCasting on double? {
   String get forceString => "${this ?? ""}";
 
   /// forcing NullableDouble type to Integer
-  int get forceInteger =>
-      int.tryParse("${this ?? "0"}".replaceAll(".", "").replaceAll(",", "")) ??
-      0;
+  int get forceInteger => this?.toInt() ?? 0;
 
   /// forcing NullableDouble type to DateTime
   DateTime get forceDateTime =>
       DateTime.tryParse("${this ?? DateTime.now()}") ?? DateTime.now();
 
   /// forcing NullableDouble type to Double
-  double get forceDouble =>
-      double.tryParse("${this ?? "0"}".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("${this ?? "0"}") ?? 0;
 
   /// forcing NullableDouble type to Boolean
   bool get forceBoolean =>
@@ -159,14 +124,13 @@ extension BooleanCasting on bool {
   String get forceString => "$this";
 
   /// forcing Boolean type to Integer
-  int get forceInteger =>
-      int.tryParse("$this".replaceAll(".", "").replaceAll(",", "")) ?? 0;
+  int get forceInteger => int.tryParse("$this") ?? 0;
 
   /// forcing Boolean type to DateTime
   DateTime get forceDateTime => DateTime.tryParse("$this") ?? DateTime.now();
 
   /// forcing Boolean type to Double
-  double get forceDouble => double.tryParse("$this".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("$this") ?? 0;
 
   /// forcing Boolean type to Boolean
   bool get forceBoolean =>
@@ -178,17 +142,14 @@ extension NullableBooleanCasting on bool? {
   String get forceString => "${this ?? ""}";
 
   /// forcing NullableBoolean type to Integer
-  int get forceInteger =>
-      int.tryParse("${this ?? "0"}".replaceAll(".", "").replaceAll(",", "")) ??
-      0;
+  int get forceInteger => int.tryParse("${this ?? "0"}") ?? 0;
 
   /// forcing NullableBoolean type to DateTime
   DateTime get forceDateTime =>
       DateTime.tryParse("${this ?? DateTime.now()}") ?? DateTime.now();
 
   /// forcing NullableBoolean type to Double
-  double get forceDouble =>
-      double.tryParse("${this ?? "0"}".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("${this ?? "0"}") ?? 0;
 
   /// forcing NullableBoolean type to Boolean
   bool get forceBoolean =>
@@ -202,14 +163,13 @@ extension DateTimeCasting on DateTime {
   String get forceString => "$this";
 
   /// forcing DateTime type to Integer
-  int get forceInteger =>
-      int.tryParse("$this".replaceAll(".", "").replaceAll(",", "")) ?? 0;
+  int get forceInteger => int.tryParse("$this") ?? 0;
 
   /// forcing DateTime type to DateTime
   DateTime get forceDateTime => DateTime.tryParse("$this") ?? DateTime.now();
 
   /// forcing DateTime type to Double
-  double get forceDouble => double.tryParse("$this".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("$this") ?? 0;
 
   /// forcing DateTime type to Boolean
   bool get forceBoolean =>
@@ -221,17 +181,14 @@ extension NullableDateTimeCasting on DateTime? {
   String get forceString => "${this ?? ""}";
 
   /// forcing NullableDateTime type to Integer
-  int get forceInteger =>
-      int.tryParse("${this ?? "0"}".replaceAll(".", "").replaceAll(",", "")) ??
-      0;
+  int get forceInteger => int.tryParse("${this ?? "0"}") ?? 0;
 
   /// forcing NullableDateTime to DateTime
   DateTime get forceDateTime =>
       DateTime.tryParse("${this ?? DateTime.now()}") ?? DateTime.now();
 
   /// forcing NullableDateTime type to Double
-  double get forceDouble =>
-      double.tryParse("${this ?? "0"}".replaceAll(",", ".")) ?? 0;
+  double get forceDouble => double.tryParse("${this ?? "0"}") ?? 0;
 
   /// forcing NullableDateTime type to Boolean
   bool get forceBoolean =>
@@ -244,17 +201,19 @@ extension NullableDateTimeCasting on DateTime? {
 String forceString(var value) => "${value ?? ""}";
 
 /// forcing dynamic type to Integer
-int forceInteger(var value) =>
-    int.tryParse("${value ?? "0"}".replaceAll(".", "").replaceAll(",", "")) ??
-    0;
+int forceInteger(var value) {
+  if (value is double) {
+    return value.toInt();
+  }
+  return int.tryParse("${value ?? "0"}") ?? 0;
+}
 
 /// forcing dynamic type to DateTime
 DateTime forceDateTime(var value) =>
     DateTime.tryParse("${value ?? DateTime.now()}") ?? DateTime.now();
 
 /// forcing dynamic type to Double
-double forceDouble(var value) =>
-    double.tryParse("${value ?? "0"}".replaceAll(",", ".")) ?? 0;
+double forceDouble(var value) => double.tryParse("${value ?? "0"}") ?? 0;
 
 /// forcing dynamic type to Boolean
 bool forceBoolean(var value) =>